Add files via upload
Core application created by yours truly. I've made a basic weather application that fetches and displays weather data for the user as well as a calendar that updates to the current day and highlights it. This should be the core and the starting point of what we are working on.